+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/table/FilterTableColumnComponentPanelTest.java
@@ -38,7 +38,7 @@ public class FilterTableColumnComponentPanelTest {
 	private final FilterTableColumn<Integer> column1 = FilterTableColumn.filterTableColumn(1);
 	private final FilterTableColumn<Integer> column2 = FilterTableColumn.filterTableColumn(2);
 
-	private final FilterTableModel.Columns<Object, Integer> columns = new FilterTableModel.Columns<>() {
+	private final FilterTableModel.TableColumns<Object, Integer> columns = new FilterTableModel.TableColumns<>() {
 		@Override
 		public List<Integer> identifiers() {
 			return asList(0, 1, 2);